Add exception handling, update Task ID type, and configure OpenAPI documentation

+++ src/main/java/com/AD/Workflow/domain/model/Task.java
@@ -11,7 +11,7 @@ public class Task {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    Long id;
+    int id;
 
     String name;
 
@@ -37,11 +37,11 @@ public class Task {
 
     public Task(){}
 
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
